server/requirements.txt
fastapi
pymongo[srv]==3.12
python-dotenv
uvicorn
pydantic
